Seperate download,upload endpoints, add flag to download
