+++ explainability-service/src/main/java/org/kie/trustyai/service/endpoints/data/UploadEndpoint.java
@@ -24,17 +24,8 @@ import org.kie.trustyai.explainability.model.SimplePrediction;
 import org.kie.trustyai.explainability.utils.IOUtils;
 import org.kie.trustyai.service.data.DataSource;
 import org.kie.trustyai.service.data.exceptions.DataframeCreateException;
-import org.kie.trustyai.service.data.metadata.Metadata;
-import org.kie.trustyai.service.data.parsers.CSVParser;
-import org.kie.trustyai.service.data.utils.DownloadUtils;
 import org.kie.trustyai.service.data.utils.UploadUtils;
-import org.kie.trustyai.service.payloads.data.download.DataRequestPayload;
-import org.kie.trustyai.service.payloads.data.download.DataResponsePayload;
-import org.kie.trustyai.service.payloads.data.download.MatchOperation;
-import org.kie.trustyai.service.payloads.data.download.RowMatcher;
 import org.kie.trustyai.service.payloads.data.upload.ModelInferJointPayload;
-import org.kie.trustyai.service.payloads.values.DataType;
-import org.kie.trustyai.service.validators.data.ValidDataDownloadRequest;
 import org.kie.trustyai.service.validators.generic.GenericValidationUtils;
 
 import jakarta.enterprise.context.ApplicationScoped;
@@ -49,100 +40,11 @@ import jakarta.ws.rs.core.Response;
 
 @ApplicationScoped
 @Path("/data")
-public class DataEndpoint {
+public class UploadEndpoint {
 
-    private static final Logger LOG = Logger.getLogger(DataEndpoint.class);
+    private static final Logger LOG = Logger.getLogger(UploadEndpoint.class);
     @Inject
     Instance<DataSource> dataSource;
-
-    CSVParser csvParser = new CSVParser();
-
-    public final static String TRUSTY_PREFIX = "trustyai.";
-
-    @POST
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    @Path("/download")
-    public Response download(@ValidDataDownloadRequest DataRequestPayload dataRequestPayload) {
-        String modelId = dataRequestPayload.getModelId();
-        Dataframe df = dataSource.get().getDataframe(modelId).copy();
-        Metadata metadata = dataSource.get().getMetadata(modelId);
-
-        for (RowMatcher rowMatcher : dataRequestPayload.getMatchAll()) {
-            if (rowMatcher.getColumnName().startsWith(DataEndpoint.TRUSTY_PREFIX)) {
-                Dataframe.InternalColumn internalColumn = Dataframe.InternalColumn.valueOf(rowMatcher.getColumnName().replace(TRUSTY_PREFIX, ""));
-                if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.BETWEEN) {
-                    df = DownloadUtils.betweenMatcherInternal(df, rowMatcher, internalColumn, false);
-                } else if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.EQUALS) {
-                    df = DownloadUtils.equalsMatcherInternal(df, rowMatcher, internalColumn, false);
-                }
-            } else {
-                int columnIndex = df.getColumnNames().indexOf(rowMatcher.getColumnName());
-                DataType columnType = DownloadUtils.getDataType(metadata, rowMatcher);
-
-                // row match
-                if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.BETWEEN) {
-                    df = DownloadUtils.betweenMatcher(df, rowMatcher, columnIndex, columnType, false);
-                } else if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.EQUALS) {
-                    df = DownloadUtils.equalsMatcher(df, rowMatcher, columnIndex, columnType, false);
-                }
-            }
-        }
-
-        for (RowMatcher rowMatcher : dataRequestPayload.getMatchNone()) {
-            if (rowMatcher.getColumnName().startsWith(DataEndpoint.TRUSTY_PREFIX)) {
-                Dataframe.InternalColumn internalColumn = Dataframe.InternalColumn.valueOf(rowMatcher.getColumnName().replace(TRUSTY_PREFIX, ""));
-                if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.BETWEEN) {
-                    df = DownloadUtils.betweenMatcherInternal(df, rowMatcher, internalColumn, true);
-                } else if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.EQUALS) {
-                    df = DownloadUtils.equalsMatcherInternal(df, rowMatcher, internalColumn, true);
-                }
-            } else {
-                int columnIndex = df.getColumnNames().indexOf(rowMatcher.getColumnName());
-                DataType columnType = DownloadUtils.getDataType(metadata, rowMatcher);
-
-                // row match
-                if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.BETWEEN) {
-                    df = DownloadUtils.betweenMatcher(df, rowMatcher, columnIndex, columnType, true);
-                } else if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.EQUALS) {
-                    df = DownloadUtils.equalsMatcher(df, rowMatcher, columnIndex, columnType, true);
-                }
-            }
-        }
-
-        Dataframe returnDF;
-        if (!dataRequestPayload.getMatchAny().isEmpty()) {
-            returnDF = df.filterByColumnValue(0, v -> false); //get null df
-            for (RowMatcher rowMatcher : dataRequestPayload.getMatchAny()) {
-                if (rowMatcher.getColumnName().startsWith(DataEndpoint.TRUSTY_PREFIX)) {
-                    Dataframe.InternalColumn internalColumn = Dataframe.InternalColumn.valueOf(rowMatcher.getColumnName().replace(TRUSTY_PREFIX, ""));
-                    if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.BETWEEN) {
-                        returnDF.addPredictions(DownloadUtils.betweenMatcherInternal(df, rowMatcher, internalColumn, false).asPredictions());
-                    } else if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.EQUALS) {
-                        returnDF.addPredictions(DownloadUtils.equalsMatcherInternal(df, rowMatcher, internalColumn, false).asPredictions());
-                    }
-                } else {
-                    int columnIndex = df.getColumnNames().indexOf(rowMatcher.getColumnName());
-                    DataType columnType = DownloadUtils.getDataType(metadata, rowMatcher);
-
-                    // row match
-                    if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.BETWEEN) {
-                        returnDF.addPredictions(DownloadUtils.betweenMatcher(df, rowMatcher, columnIndex, columnType, false).asPredictions());
-                    } else if (MatchOperation.valueOf(rowMatcher.getOperation()) == MatchOperation.EQUALS) {
-                        returnDF.addPredictions(DownloadUtils.equalsMatcher(df, rowMatcher, columnIndex, columnType, false).asPredictions());
-                    }
-                }
-            }
-        } else {
-            returnDF = df;
-        }
-
-        DataResponsePayload dataResponsePayload = new DataResponsePayload();
-        dataResponsePayload.setDataCSV(csvParser.convertToString(returnDF, true, true));
-
-        return Response.ok().entity(dataResponsePayload).build();
-
-    }
 
     // handle the ground truth logic
     // if successful, return Pair of dataframe, response message
