+++ explainability-service/src/main/resources/application.properties
@@ -49,12 +49,24 @@ quarkus.banner.path=banner.txt
 quarkus.http.handle-100-continue-automatically=true
 
 # Feature flags
+# default/dev endpoints
+endpoints.fairness=enable
+endpoints.drift=enable
+endpoints.explainers.local=enable
+endpoints.explainers.global=enable
+endpoints.data.download=disable
+
+# odh endpoints
 %odh.endpoints.fairness=enable
 %odh.endpoints.drift=enable
 %odh.endpoints.explainers.local=enable
 %odh.endpoints.explainers.global=enable
 
+# rhoai endpoints
 %rhoai.endpoints.fairness=disable
 %rhoai.endpoints.drift=enable
 %rhoai.endpoints.explainers.local=disable
 %rhoai.endpoints.explainers.global=disable
+
+# test profile endpoint overrides
+%test.endpoints.data.download=enable
